Convert cost of upgrades to long value for higher economy servers.

+++ src/main/java/com/massivecraft/factions/FPlayer.java
@@ -255,9 +255,9 @@ public interface FPlayer extends EconomyParticipator {
 
     int getDeaths();
 
-    boolean takeMoney(int amt);
+    boolean takeMoney(long amt);
 
-    boolean hasMoney(int amt);
+    boolean hasMoney(long amt);
 
     //Fplayer specific friendly fire.
     boolean hasFriendlyFire();
